# Conflicts:
#	continew-webapi/src/main/java/top/continew/admin/controller/schedule/JobController.java
#	continew-webapi/src/main/java/top/continew/admin/controller/system/DictItemController.java
#	continew-webapi/src/main/java/top/continew/admin/controller/system/FileController.java
#	continew-webapi/src/main/resources/db/changelog/mysql/main_data.sql
#	continew-webapi/src/main/resources/db/changelog/postgresql/main_data.sql

+++ continew-module-system/src/main/java/top/continew/admin/system/service/impl/DictItemServiceImpl.java
@@ -16,6 +16,7 @@
 
 package top.continew.admin.system.service.impl;
 
+import cn.hutool.core.collection.CollUtil;
 import cn.hutool.core.util.ClassUtil;
 import cn.hutool.core.util.StrUtil;
 import com.alicp.jetcache.anno.Cached;
@@ -77,6 +78,9 @@ public class DictItemServiceImpl extends BaseServiceImpl<DictItemMapper, DictIte
 
     @Override
     public void deleteByDictIds(List<Long> dictIds) {
+        if (CollUtil.isEmpty(dictIds)) {
+            return;
+        }
         baseMapper.lambdaUpdate().in(DictItemDO::getDictId, dictIds).remove();
         RedisUtils.deleteByPattern(CacheConstants.DICT_KEY_PREFIX + StringConstants.ASTERISK);
     }
